+++ app/models/concerns/ruby_llm/evals/executable.rb
@@ -4,6 +4,10 @@ module RubyLLM
       extend ActiveSupport::Concern
 
       def execute(variables: {}, files: [])
+        to_chat(variables:, files:).complete
+      end
+
+      def to_chat(variables: {}, files: [])
         chat = RubyLLM.chat(model: model, provider: provider)
 
         chat.with_instructions Liquid::Template.parse(instructions).render(variables) if instructions.present?
@@ -25,7 +29,11 @@ module RubyLLM
         chat.with_thinking(effort: thinking_effort) if thinking_effort.present?
         chat.with_thinking(budget: thinking_budget) if thinking_budget.present?
 
-        chat.ask Liquid::Template.parse(message).render(variables), with: files
+        rendered_message = Liquid::Template.parse(message).render(variables)
+        content = files.present? ? RubyLLM::Content.new(rendered_message, files) : rendered_message
+        chat.add_message role: :user, content: content
+
+        chat
       end
     end
   end